AvtoInfo-master/pyproject.toml
[tool.poetry]
name = "avtoinfo"
version = "0.1.0"
description = ""
authors = ["Lev <lafitel31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
